Remove some old comments
